+++ src/lib.rs
@@ -29,6 +29,7 @@ pub mod api;
 #[doc(hidden)]
 pub mod builtin_tools;
 pub mod config;
+pub mod dotenv;
 #[doc(hidden)]
 pub mod io_dispatcher;
 pub mod io_handler;